requirements.txt
certifi==2021.10.8
cycler==0.11.0
et-xmlfile==1.1.0
fonttools==4.29.0
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.1
openpyxl==3.0.9
packaging==21.3
pandas==1.4.1
Pillow==9.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
SciencePlots==1.0.9
scipy==1.7.3
six==1.16.0
